refactor : S3 관련 Service 네이밍 수정
