--- backend/src/test/java/com/mapbefine/mapbefine/FileFixture.java
@@ -1,11 +1,0 @@
-package com.mapbefine.mapbefine;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public class FileFixture {
-
-    public static MultipartFile createFile() {
-        return new StubFile();
-    }
-
-}
