+++ backend/src/test/java/com/mapbefine/mapbefine/image/FileFixture.java
@@ -1,0 +1,17 @@
+package com.mapbefine.mapbefine.image;
+
+import org.springframework.mock.web.MockMultipartFile;
+import org.springframework.web.multipart.MultipartFile;
+
+public class FileFixture {
+
+    public static MultipartFile createFile() {
+        return new MockMultipartFile(
+                "test",
+                "test.png",
+                "img/png",
+                "image".getBytes()
+        );
+    }
+
+}
